feat: Added embedded and iiif samples to the project.
These are included in the IArchives BookReader distribution on github. We do not use, but might as well have on hand as examples.  The github repo also includes a sample for the IArchives Plato dialogs; this appears more or less identical to the simple demo, with a http call to retrieve data from IArchives.

Issues: AC-1179